Refactor initialisation part but need to review
I have cleaned up the code and found out about the exiting. Not sure
about if this is way too restricted for shell program. But my argument
is that you need to initalise the shell program properly. so if there
are any issues, restart the program until it is successful!